fix checkpoint

+++ src/runtime/streaming/operators/joins/join_instance.rs
@@ -360,16 +360,16 @@ impl Operator for InstantJoinOperator {
         self.state_store
             .as_ref()
             .unwrap()
-            .prepare_checkpoint_epoch(barrier.epoch as u64)
+            .prepare_checkpoint_epoch(barrier.epoch)
             .map_err(|e| anyhow!("Snapshot failed: {e}"))?;
         Ok(())
     }
 
-    async fn commit_checkpoint(&mut self, epoch: u32, _ctx: &mut TaskContext) -> Result<()> {
+    async fn commit_checkpoint(&mut self, epoch: u64, _ctx: &mut TaskContext) -> Result<()> {
         self.state_store
             .as_ref()
             .unwrap()
-            .commit_checkpoint_epoch(epoch as u64)
+            .commit_checkpoint_epoch(epoch)
             .map_err(|e| anyhow!("Commit checkpoint failed: {e}"))?;
         Ok(())
     }
