+++ src/runtime/streaming/operators/joins/join_with_expiration.rs
@@ -314,18 +314,18 @@ impl Operator for JoinWithExpirationOperator {
             .expect("State store not initialized");
 
         store
-            .prepare_checkpoint_epoch(barrier.epoch as u64)
+            .prepare_checkpoint_epoch(barrier.epoch)
             .map_err(|e| anyhow!("Snapshot failed: {e}"))?;
 
         info!(epoch = barrier.epoch, "Join Operator snapshotted state.");
         Ok(())
     }
 
-    async fn commit_checkpoint(&mut self, epoch: u32, _ctx: &mut TaskContext) -> Result<()> {
+    async fn commit_checkpoint(&mut self, epoch: u64, _ctx: &mut TaskContext) -> Result<()> {
         self.state_store
             .as_ref()
             .expect("State store not initialized")
-            .commit_checkpoint_epoch(epoch as u64)
+            .commit_checkpoint_epoch(epoch)
             .map_err(|e| anyhow!("Commit checkpoint failed: {e}"))?;
         Ok(())
     }
