+++ src/runtime/streaming/operators/sink/iceberg/mod.rs
@@ -132,7 +132,7 @@ impl IcebergSinkOperator {
         })
     }
 
-    async fn flush_epoch(&mut self, epoch: u32, subtask_idx: usize) -> Result<()> {
+    async fn flush_epoch(&mut self, epoch: u64, subtask_idx: usize) -> Result<()> {
         if self.pending.is_empty() {
             return Ok(());
         }
@@ -263,7 +263,7 @@ impl Operator for IcebergSinkOperator {
                 table = %self.table_name,
                 "flushing remaining iceberg sink batches on close"
             );
-            self.flush_epoch(u32::MAX, ctx.subtask_index as usize)
+            self.flush_epoch(u64::MAX, ctx.subtask_index as usize)
                 .await?;
         }
         Ok(vec![])
